+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Added
 * `shopify/prefer-singular-enums` ([#132](https://github.com/Shopify/eslint-plugin-shopify/pull/132))
 * `shopify/react-no-multiple-render-methods` ([#134](https://github.com/Shopify/eslint-plugin-shopify/pull/134))
+* `shopify/prefer-explicit-local-imports` ([#149](https://github.com/Shopify/eslint-plugin-shopify/pull/149))
 
 
 ## [23.1.0] - 2018-08-02s